requirements.txt
## Core Dependencies
# Python==3.10

## API & Web Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

## Data Processing
numpy==1.24.3
pandas==2.0.3
polars==0.19.12
scikit-learn==1.3.2

## Machine Learning
tensorflow==2.14.0
tensorflow-io==0.33.0
torch==2.1.1
xgboost==2.0.1
lightgbm==4.1.0
catboost==1.2.2

## Time Series Forecasting
statsmodels==0.14.0
prophet==1.1.5
pmdarima==2.0.4

## Optimization
pulp==2.7.0
ortools==9.7.2996
cvxpy==1.3.2

## Model Management & Serving
mlflow==2.8.1
joblib==1.3.2

## Monitoring & Logging
python-json-logger==2.0.7
prometheus-client==0.19.0

## Utilities
python-dotenv==1.0.0
requests==2.31.0
redis==5.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1

## Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.12.0
flake8==6.1.0

## Data Validation
whylogs==1.1.5

## GPU Support (optional, remove if CPU-only)
# cuda-python==12.3.0
# cupy==12.3.0

## Docker & Deployment
gunicorn==21.2.0
